requirements.txt
Django==1.5.1
PyYAML==3.10
Pygments==1.6
djangorestframework==2.3.8